requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
google-genai==0.7.0
pypdf==4.3.1
python-docx==1.1.2
certifi==2024.8.30
